+++ sdk-e2e/src/test/kotlin/org/aerogear/sdkE2ETests/init.kt
@@ -1,11 +1,11 @@
-package org.aerogear.sdk_e2e_tests
+package org.aerogear.sdkE2ETests
 
 import org.aerogear.akow.dsl.base.appium
 import org.aerogear.akow.example.test.pageobjects.android.MainAndroidScreen
-import org.aerogear.sdk_e2e_tests.pageobjects.android.AuthStartAndroidScreen
-import org.aerogear.sdk_e2e_tests.pageobjects.android.AuthenticatedAndroidScreen
-import org.aerogear.sdk_e2e_tests.pageobjects.android.KeycloakLoginAndroidScreen
-import org.aerogear.sdk_e2e_tests.pageobjects.unimportant.SetPINDialogAndroidScreen
+import org.aerogear.sdkE2ETests.pageobjects.android.AuthStartAndroidScreen
+import org.aerogear.sdkE2ETests.pageobjects.android.AuthenticatedAndroidScreen
+import org.aerogear.sdkE2ETests.pageobjects.android.KeycloakLoginAndroidScreen
+import org.aerogear.sdkE2ETests.pageobjects.unimportant.SetPINDialogAndroidScreen
 
 val appium = appium(BuildConfig.APPIUM_SERVER_URL) {
     applications {
