+++ sdk-e2e/src/test/kotlin/org/aerogear/sdkE2ETests/pageobjects/android/AuthStartAndroidScreen.kt
@@ -1,7 +1,7 @@
-package org.aerogear.sdk_e2e_tests.pageobjects.android
+package org.aerogear.sdkE2ETests.pageobjects.android
 
 import org.aerogear.akow.pageobject.AndroidScreen
-import org.aerogear.sdk_e2e_tests.pageobjects.AuthStartScreen
+import org.aerogear.sdkE2ETests.pageobjects.AuthStartScreen
 
 /**
  * Authentication start screen for Android
